+++ src/grids/cvt/ellipse_cvt.cpp
@@ -122,15 +122,15 @@ void EllipseCVT::fillBoundaryPoints(int dim_num, int nb_nodes, int *seed, double
 
     // total nb points used to compute Voronoi mesh. 
     // Only (nb_interior_pts-nb_bnd) will be able to move freely
-    unsigned int tot_nb_pts = nb_nodes;
+    int tot_nb_pts = nb_nodes;
     // number of boundary points, automatically calculated
     printf("tot_nb_pts= %d\n", tot_nb_pts);
     printf("domain integral = %f\n", dom_intg);
     printf("boundary integral = %f\n", bnd_intg);
 
-    unsigned int nb_computed_bnd = bnd_intg * sqrt(tot_nb_pts / dom_intg);
-    unsigned int nb_bnd_1 = 1. + 16. * tot_nb_pts * dom_intg / (bnd_intg * bnd_intg);
-    nb_bnd_1 = -bnd_intg * bnd_intg / (4. * dom_intg) * (1. - sqrt(nb_bnd_1));
+    int nb_computed_bnd = (int)(bnd_intg * sqrt(tot_nb_pts / dom_intg));
+    int nb_bnd_1 = (int)(1. + 16. * tot_nb_pts * dom_intg / (bnd_intg * bnd_intg));
+    nb_bnd_1 = (int)(-bnd_intg * bnd_intg / (4. * dom_intg) * (1. - sqrt(nb_bnd_1)));
     nb_computed_bnd = nb_bnd_1; // more accurate formula
     printf("calculated nb boundary pts: %d\n", nb_computed_bnd);
     printf("improved nb boundary pts: %d\n", nb_bnd_1);
